requirements.txt
numpy>=1.24.0
pyyaml>=6.0
msgpack>=1.0.0
watchdog>=3.0.0
